+++ src/Grid/Displayers/BelongsTo.php
@@ -3,90 +3,21 @@
 namespace Encore\Admin\Grid\Displayers;
 
 use Encore\Admin\Admin;
+use Encore\Admin\Grid\Selectable;
 
 class BelongsTo extends AbstractDisplayer
 {
-    use BelongsToRelation;
-
     /**
-     * @return $this
+     * @param int $multiple
+     *
+     * @return string
      */
-    public function addScript()
+    protected function getLoadUrl($selectable, $multiple = 0)
     {
-        $script = <<<SCRIPT
-;(function () {
-    var modal = $('#{$this->modalID}');
-    var related = null;
-    var selected = {};
+        $selectable = str_replace('\\', '_', $selectable);
+        $args = [$multiple];
 
-    var load = function (url) {
-        $.get(url, function (data) {
-            modal.find('.modal-body').html(data);
-            modal.find('.select').iCheck({
-                radioClass:'iradio_minimal-blue',
-                checkboxClass:'icheckbox_minimal-blue'
-            });
-            modal.find('.box-header:first').hide();
-
-            modal.find('input.select').each(function (index,    el) {
-                if ($(el).val() == selected.id) {
-                    $(el).iCheck('toggle');
-                }
-            });
-        });
-    };
-
-    var update = function (callback) {
-        $.ajax({
-            url: "{$this->getResource()}/" + related.attr('key'),
-            type: "POST",
-            data: {
-                {$this->columnName}: selected.id,
-                _token: LA.token,
-                _method: 'PUT'
-            },
-            success: function (data) {
-                callback(data);
-            }
-        });
-    };
-
-    modal.on('show.bs.modal', function (e) {
-        related = $(e.relatedTarget);
-        selected.id = related.data('val');
-        load("{$this->getLoadUrl()}");
-    }).on('click', '.page-item a, .filter-box a', function (e) {
-        load($(this).attr('href'));
-        e.preventDefault();
-    }).on('click', 'tr', function (e) {
-        $(this).find('input.select').iCheck('toggle');
-        e.preventDefault();
-    }).on('submit', '.box-header form', function (e) {
-        load($(this).attr('action')+'&'+$(this).serialize());
-        return false;
-    }).on('ifChecked', 'input.select', function (e) {
-        selected.id = $(this).val();
-    }).find('.modal-footer .submit').click(function () {
-        update(function (data) {
-            related.data('val', selected);
-            related.find('.text').html(data.display["{$this->columnName}"]);
-            related.find('a').toggleClass('text-green text-muted');
-
-            setTimeout(function () {
-                related.find('a').toggleClass('text-green text-muted');
-            }, 2000);
-
-            modal.modal('toggle');
-
-            toastr.success(data.message);
-        });
-    });
-})();
-SCRIPT;
-
-        Admin::script($script);
-
-        return $this;
+        return route('admin.handle-selectable', compact('selectable', 'args'));
     }
 
     /**
@@ -95,5 +26,31 @@ SCRIPT;
     protected function getOriginalData()
     {
         return $this->getColumn()->getOriginal();
+    }
+
+    /**
+     * @param string $selectable
+     * @param string $column
+     *
+     * @return string
+     */
+    public function display($selectable = null, $column = '')
+    {
+        if (!class_exists($selectable) || !is_subclass_of($selectable, Selectable::class)) {
+            throw new \InvalidArgumentException(
+                "[Class [{$selectable}] must be a sub class of Encore\Admin\Grid\Selectable"
+            );
+        }
+
+        return Admin::component('admin::grid.inline-edit.belongsto', [
+            'modal'     => sprintf('modal-grid-selector-%s', $this->getClassName()),
+            'key'       => $this->getKey(),
+            'original'  => $this->getOriginalData(),
+            'value'     => $this->getValue(),
+            'resource'  => $this->getResource(),
+            'name'      => $column ?: $this->getName(),
+            'relation'  => get_called_class(),
+            'url'       => $this->getLoadUrl($selectable, get_called_class() == BelongsToMany::class),
+        ]);
     }
 }
