+++ src/Grid/Displayers/BelongsToMany.php
@@ -6,109 +6,14 @@ use Encore\Admin\Admin;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany as Relation;
 use Illuminate\Support\Arr;
 
-class BelongsToMany extends AbstractDisplayer
+class BelongsToMany extends BelongsTo
 {
-    use BelongsToRelation;
-
     /**
      * Other key for many-to-many relation.
      *
      * @var string
      */
     protected static $otherKey = [];
-
-    /**
-     * @return $this
-     */
-    public function addScript()
-    {
-        $script = <<<SCRIPT
-;(function () {
-    var modal = $('#{$this->modalID}');
-    var related = null;
-    var selected = [];
-
-    var load = function (url) {
-        $.get(url, function (data) {
-            modal.find('.modal-body').html(data);
-            modal.find('.select').iCheck({
-                radioClass:'iradio_minimal-blue',
-                checkboxClass:'icheckbox_minimal-blue'
-            });
-            modal.find('.box-header:first').hide();
-
-            modal.find('input.select').each(function (index, el) {
-                if ($.inArray($(el).val().toString(), selected) >=0 ) {
-                    $(el).iCheck('toggle');
-                }
-            });
-        });
-    };
-
-    var update = function (callback) {
-        $.ajax({
-            url: "{$this->getResource()}/" + related.attr('key'),
-            type: "POST",
-            data: {
-                {$this->columnName}: selected,
-                _token: LA.token,
-                _method: 'PUT'
-            },
-            success: function (data) {
-                callback(data);
-            }
-        });
-    };
-
-    modal.on('show.bs.modal', function (e) {
-        related = $(e.relatedTarget);
-        selected = related.data('val').map(function (value) {
-            return value.toString();
-        });
-
-        load("{$this->getLoadUrl(1)}");
-    }).on('click', '.page-item a, .filter-box a', function (e) {
-        load($(this).attr('href'));
-        e.preventDefault();
-    }).on('click', 'tr', function (e) {
-        $(this).find('input.select').iCheck('toggle');
-        e.preventDefault();
-    }).on('submit', '.box-header form', function (e) {
-        load($(this).attr('action')+'&'+$(this).serialize());
-        e.preventDefault();
-    }).on('ifChecked', 'input.select', function (e) {
-        var val = $(this).val().toString();
-        if (selected.indexOf(val) < 0) {
-            selected.push(val);
-        }
-    }).on('ifUnchecked', 'input.select', function (e) {
-           var val = $(this).val().toString();
-           var index = selected.indexOf(val);
-            if (index !== -1) {
-                selected.splice(index, 1);
-            }
-    }).find('.modal-footer .submit').click(function () {
-        update(function (data) {
-            related.data('val', selected);
-            related.find('.text').html(data.display["{$this->columnName}"]);
-            related.find('a').toggleClass('text-green text-muted');
-
-            setTimeout(function () {
-                related.find('a').toggleClass('text-green text-muted');
-            }, 2000);
-
-            modal.modal('toggle');
-
-            toastr.success(data.message);
-        });
-    });
-})();
-SCRIPT;
-
-        Admin::script($script);
-
-        return $this;
-    }
 
     /**
      * Get other key for this many-to-many relation.
